+++ src/cavalion-blocks/Node<>/fs.js
@@ -1,2 +1,0 @@
-// Just an alias
-["Node<devtools/Resources>"]