+++ src/cavalion-blocks/ide/Main.workspaces.js
@@ -1,7 +1,0 @@
-define([
-	["#tabs", [
-		["Tab<Project>", "code"],
-		["Tab<Project>", "vcl"],
-		["Tab<Project>", "blocks"]
-	]]
-]);