+++ src/cavalion-blocks/ide/Searchbar.js
@@ -1,7 +1,0 @@
-"use strict";
-
-var locale = window.locale;
-
-["vcl-ui:Bar", { classes: "no-border" }, [
-	["vcl-ui:Input", "search", { placeholder: locale("Search.placeholder") }]
-]];