+++ src/cavalion-blocks/ide/Topbar.js
@@ -1,3 +1,0 @@
-"use strict";
-
-["vcl-ui:Bar", { css: "background-color: red;"}, []];