+++ src/vcl-comps/devtools/Workspace.js
@@ -200,6 +200,8 @@ $(["ui/Form"], {
                         evt.formUri = "devtools/Editor<blocks>";
                     // } else if(path.indexOf("pages") !== -1)  {
                     // 	evt.formUri = "devtools/Editor<page>";
+    			    } else if(evt.resource.uri.indexOf("/var/log/") !== -1) {
+    			    	evt.formUri = "devtools/Editor<var/log>";
                     } else {
     			        evt.formUri = String.format("devtools/Editor<%s>", ext);
                     }